+++ packages/sources/src/components/justify-form-groups/justify-form-groups.models.ts
@@ -1,0 +1,7 @@
+import { Button } from '../button/button.models';
+import { FormGroup } from '../form/models/form-group.model';
+
+export interface JustifyFormGroups<TemplateFnReturnType> {
+  formGroups: FormGroup<TemplateFnReturnType>[];
+  buttons: Button[];
+}